refactor(string-adder): refactor codes in string-adder
