+++ src/test/kotlin/adder/model/NumberTest.kt
@@ -1,6 +1,5 @@
-package adder
+package adder.model
 
-import adder.model.Number
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
