+++ src/test/kotlin/adder/model/NumbersTest.kt
@@ -1,7 +1,5 @@
-package adder
+package adder.model
 
-import adder.model.Number
-import adder.model.Numbers
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
